[WIP] Decoupling utils.py (eventually removed)
