+++ caterva2/api_utils.py
@@ -11,7 +11,6 @@ import pickle
 # Requirements
 import httpx
 
-# To remove
 
 
 def parse_slice(string):
