+++ caterva2/clients/cli.py
@@ -16,7 +16,9 @@ import httpx
 import rich
 
 # Project
-from caterva2 import utils, api_utils, models
+from caterva2 import api_utils, models
+from caterva2.services import srv_utils
+from caterva2.clients import cli_utils
 
 
 def handle_errors(func):
@@ -109,7 +111,7 @@ def cmd_info(args):
 def cmd_show(args):
     # Download
     dataset, params = args.dataset
-    array, schunk = utils.download(args.host, dataset, params, verbose=True)
+    array, schunk = srv_utils.download(args.host, dataset, params, verbose=True)
 
     # Display
     if array is None:
@@ -138,7 +140,7 @@ def cmd_download(args):
         localpath = pathlib.Path(f'{localpath}[{slice}]{suffix}')
 
     # Download
-    array, schunk = utils.download(args.host, dataset, params, localpath=localpath, verbose=True)
+    array, schunk = srv_utils.download(args.host, dataset, params, localpath=localpath, verbose=True)
     if suffix not in {'.b2frame', '.b2nd'}:
         with open(localpath, 'wb') as f:
             data = schunk[:]
@@ -147,7 +149,7 @@ def cmd_download(args):
     print(f'Dataset saved to {localpath}')
 
 if __name__ == '__main__':
-    parser = utils.get_parser()
+    parser = cli_utils.get_parser()
     parser.add_argument('--host', default='localhost:8002')
     subparsers = parser.add_subparsers(required=True)
 
@@ -201,5 +203,5 @@ if __name__ == '__main__':
     subparser.set_defaults(func=cmd_download)
 
     # Go
-    args = utils.run_parser(parser)
+    args = cli_utils.run_parser(parser)
     args.func(args)
