+++ caterva2/clients/cli_utils.py
@@ -1,0 +1,40 @@
+###############################################################################
+# Caterva2 - On demand access to remote Blosc2 data repositories
+#
+# Copyright (c) 2023 The Blosc Developers <blosc@blosc.org>
+# https://www.blosc.org
+# License: GNU Affero General Public License v3.0
+# See LICENSE.txt for details about copyright and rights to use.
+###############################################################################
+
+import argparse
+import logging
+
+
+#
+# Command line helpers
+#
+def socket_type(string):
+    host, port = string.split(':')
+    port = int(port)
+    return (host, port)
+
+
+def get_parser(broker=None, http=None):
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--loglevel', default='warning')
+    if broker:
+        parser.add_argument('--broker', default=broker)
+    if http:
+        parser.add_argument('--http', default=http, type=socket_type)
+    return parser
+
+
+def run_parser(parser):
+    args = parser.parse_args()
+
+    # Logging
+    loglevel = args.loglevel.upper()
+    logging.basicConfig(level=loglevel)
+
+    return args
